refactor(site): convert AgentChatInput tests to Storybook stories
Replace AgentChatInput.test.tsx (RTL/vitest) with
AgentChatInput.stories.tsx using Storybook play functions.

Stories cover all original test scenarios:
- Default: baseline render
- DisablesSendUntilInput: send button disabled until text entered
- SendsAndClearsInput: sends message and clears textarea
- DisabledInput: send stays disabled when isDisabled is true
- NoModelOptions: send stays disabled when hasModelOptions is false
- LoadingSpinner: shows animate-spin while loading
